stage2/00-copies-and-fills/02-run.sh: Fix re-runs of script
Check that file exist before move.

+++ stage2/00-copies-and-fills/02-run.sh
@@ -1,3 +1,6 @@
 #!/bin/bash -e
 
-mv "${ROOTFS_DIR}/etc/ld.so.preload" "${ROOTFS_DIR}/etc/ld.so.preload.disabled"
+if [ -f "${ROOTFS_DIR}/etc/ld.so.preload" ]; then
+   mv "${ROOTFS_DIR}/etc/ld.so.preload" "${ROOTFS_DIR}/etc/ld.so.preload.disabled"
+fi
+
